+++ backlib/internal/backports/py311/tomllib/__init__.py
@@ -4,7 +4,7 @@ import sys
 if sys.version_info >= (3, 11):
     from tomllib import TOMLDecodeError, load, loads
 else:
-    from backlib.py313.internal.backports.tomllib.internal.tomllib import (
+    from backlib.internal.backports.py311.tomllib.internal.tomllib import (
         TOMLDecodeError,
         load,
         loads,