+++ backlib/internal/backports/py311/tomllib/internal/cpython/parser.py
@@ -9,7 +9,7 @@ import string
 from types import MappingProxyType
 from typing import TYPE_CHECKING, Any, NamedTuple
 
-from backlib.py313.internal.backports.tomllib.internal.cpython.re import (
+from backlib.internal.backports.py311.tomllib.internal.cpython.re import (
     RE_DATETIME,
     RE_LOCALTIME,
     RE_NUMBER,
@@ -22,8 +22,8 @@ from backlib.py313.internal.backports.tomllib.internal.cpython.re import (
 if TYPE_CHECKING:
     from collections.abc import Iterable
 
+    from backlib.internal.backports.py311.tomllib.internal.cpython.types import Key, ParseFloat, Pos
     from backlib.internal.typing import SupportsRead
-    from backlib.py313.internal.backports.tomllib.internal.cpython.types import Key, ParseFloat, Pos
 
 
 ASCII_CTRL = frozenset(chr(i) for i in range(32)) | frozenset(chr(127))