+++ backlib/internal/backports/py311/tomllib/internal/tomllib.py
@@ -1,9 +1,9 @@
-import backlib.py313.internal.backports.tomllib.internal.cpython as tomllib
+import backlib.internal.backports.py311.tomllib.internal.cpython as backport
 
 
 __all__: list[str] = ["TOMLDecodeError", "load", "loads"]
 
-__backlib__: str = "backlib.py313.tomllib"
+__backlib__: str = "backlib.py311.tomllib"
 
 
 # ---
@@ -11,11 +11,11 @@ __backlib__: str = "backlib.py313.tomllib"
 # Explain: Added in Python 3.11.
 # ---
 
-TOMLDecodeError = tomllib.TOMLDecodeError
+TOMLDecodeError = backport.TOMLDecodeError
 TOMLDecodeError.__module__ = __backlib__
 
-load = tomllib.load
+load = backport.load
 load.__module__ = __backlib__
 
-loads = tomllib.loads
+loads = backport.loads
 loads.__module__ = __backlib__