+++ backlib/internal/backports/py313/tomllib.py
@@ -1,0 +1,4 @@
+from backlib.internal.backports.py312.tomllib import TOMLDecodeError, load, loads
+
+
+__all__: list[str] = ["TOMLDecodeError", "load", "loads"]