+++ backlib/py311/tomllib.py
@@ -1,0 +1,4 @@
+from backlib.internal.backports.py311.tomllib import TOMLDecodeError, load, loads
+
+
+__all__: list[str] = ["TOMLDecodeError", "load", "loads"]