+++ backlib/py313/tomllib.py
@@ -1,4 +1,4 @@
-from backlib.py313.internal.backports.tomllib import TOMLDecodeError, load, loads
+from backlib.internal.backports.py313.tomllib import TOMLDecodeError, load, loads
 
 
 __all__: list[str] = ["TOMLDecodeError", "load", "loads"]